131: separate update and delete functionality to widgets
